+++ src/main/java/insane96mcp/insanesurvivaloverhaul/module/combat/regeneratingabsorption/RegeneratingAbsorptionClient.java
@@ -23,12 +23,6 @@ public class RegeneratingAbsorptionClient {
         ResourceLocation aboveOverlay = VanillaGuiLayers.PLAYER_HEALTH;
         if (ModList.get().isLoaded("stamina"))
             aboveOverlay = ResourceLocation.parse("stamina:stamina_overlay");
-        if (RegeneratingAbsorption.renderOnTheRight) {
-            //if (ModList.get().isLoaded("nohunger") && NoHungerIntegration.doesRenderArmorAtHunger())
-            //    aboveOverlay = InsaneSO.location("armor");
-            //else
-                aboveOverlay = VanillaGuiLayers.FOOD_LEVEL;
-        }
         Minecraft mc = Minecraft.getInstance();
         Gui gui = mc.gui;
         event.registerAbove(aboveOverlay, InsaneSO.id("regenerating_absorption"), (guiGraphics, partialTicks) -> {
@@ -50,8 +44,8 @@ public class RegeneratingAbsorptionClient {
         mc.getProfiler().push("regen_absorption");
 
         RenderSystem.enableBlend();
-        int left = width / 2 + (!RegeneratingAbsorption.renderOnTheRight ? -91 : 82);
-        int top = height - (!RegeneratingAbsorption.renderOnTheRight ? gui.leftHeight : gui.rightHeight);
+        int left = width / 2 - 91;
+        int top = height - gui.leftHeight;
 
         int absorption = Mth.ceil(ModNBTData.get(mc.player, RegeneratingAbsorption.REGEN_ABSORPTION_TAG, Float.class));
         boolean highlight = absorptionBlinkTime > (long) gui.getGuiTicks() && (absorptionBlinkTime - (long) gui.getGuiTicks()) / 3L % 2L == 1L;
@@ -85,28 +79,19 @@ public class RegeneratingAbsorptionClient {
                 ClientUtils.setRenderColor(1, 0, 0, 1f);
             //ClientUtils.blitVericallyMirrored(GUI_ICONS, guiGraphics, left, top, 9, v, 9, 9, 18, 18);
             int u = i % 2 == 0 ? 0 : 9;
-            if (!RegeneratingAbsorption.renderOnTheRight)
-                guiGraphics.blit(GUI_ICONS, left, top, u, v, 9, 9, 18, 18);
-            else
-                ClientUtils.blitVerticallyMirrored(GUI_ICONS, guiGraphics,left, top, u, v, 9, 9, 18, 18);
+            guiGraphics.blit(GUI_ICONS, left, top, u, v, 9, 9, 18, 18);
             if (i % 20 == 0 && i != displayAbsorption) {
-                left = width / 2 + (!RegeneratingAbsorption.renderOnTheRight ? -91 : 82);
+                left = width / 2 - 91;
                 top -= 10;
-                if (!RegeneratingAbsorption.renderOnTheRight)
-                    gui.leftHeight += 10;
-                else
-                    gui.rightHeight += 10;
+                gui.leftHeight += 10;
             }
             else if (i % 2 == 0)
-                left += RegeneratingAbsorption.renderOnTheRight ? -8 : 8;
+                left -= 8;
             if (i > absorption)
                 ClientUtils.resetRenderColor();
         }
         if (displayAbsorption > 0)
-            if (!RegeneratingAbsorption.renderOnTheRight)
-                gui.leftHeight += 10;
-            else
-                gui.rightHeight += 10;
+            gui.leftHeight += 10;
 
         RenderSystem.disableBlend();
         mc.getProfiler().pop();